+++ common/utilities.py
@@ -1,0 +1,18 @@
+def bitand(x: int, y: int) -> int:
+    return x & y
+
+
+def bitor(x: int, y: int) -> int:
+    return x | y
+
+
+def lshift(x: int, y: int) -> int:
+    return x << y
+
+
+def rshift(x: int, y: int) -> int:
+    return x >> y
+
+
+def bitnot(x: int) -> int:
+    return x ^ 65535
